bug(password): reset password
- create reducer for resetting password
- create action for resetting password
- write task for reset password
- create container component for reset password
- create presentational component for reset password

[Deliver #166284492]
