+++ src/components/ResetPassword.jsx
@@ -1,115 +1,87 @@
-/* eslint-disable react/prop-types */
-/* eslint-disable react/jsx-no-bind */
-import React, { useState, useRef } from 'react';
+/* eslint-disable import/no-named-as-default */
+/* eslint-disable no-unused-expressions */
+import React, { Component } from 'react';
+import jwtDecode from 'jwt-decode';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
-import { ClipLoader } from 'react-spinners';
 import { createStructuredSelector } from 'reselect';
+import { connect } from 'react-redux';
+import { toast } from 'react-toastify';
 import * as selectors from '../selectors/resetPassword';
-import Modal from '../views/Modal';
-import passwordResetRequest from '../actions/resetPassword';
-import modalAction from '../actions/modal';
-
+import { resetPassword } from '../actions/resetPasswordAction';
+import ResetPasswordForm from '../views/ResetPasswordForm';
 
 /**
  * Form for resetting password
+ * @class UpdateUserPassword
+ * @extends {Component}
  * @param {object} props Component props
- * @param {function} props.passwordResetRequest action creator function
+ * @param {function} props.updatePassword action creator function
+ * @returns {JSX}
  */
-function ResetPassword(props) {
-  /** local state for handling email gotten from the input field */
-  const [userEmail, setUserEmail] = useState({
-    email: null,
-  });
+export class ResetPassword extends Component {
+  // local state for handling email gotten from the input field
+  state = {
+    password: null,
+    confirmPassword: null,
+  }
 
-  /** local state for adding a new css class to a div */
-  const [inputIsValid, setInputIsValid] = useState('');
-
-  /** mutable object for an input */
-  const input = useRef(null);
-
-  /** passing global state loading as props into variables */
-  const { message, isLoading } = props;
-
-  /** this is a function to update the email state */
-  const updateInput = (event) => {
-    setInputIsValid('validated');
-    setUserEmail({ ...userEmail, [event.target.name]: event.target.value });
+  // this is a function to update the password and confirmPassword state
+  updateInput = (event) => {
+    this.setState({
+      [event.target.name]: event.target.value
+    });
   };
 
-  /** A function to clear the input field */
-  const clearInputs = () => {
-    input.current.value = null;
-    setUserEmail({ ...userEmail, email: null });
-  };
-
-  /** A function that triggers the post request action */
-  const handleSubmit = async (event) => {
+  // A function that triggers the post request action
+  handleSubmit = async (event) => {
     event.preventDefault();
-    /** calling the action creator to update the store */
-    props.passwordResetRequest(userEmail.email);
-  };
-
-  /** A function that ensures that when modal is closed, content is lost */
-  const clearModalState = (content) => {
-    /** action to clear response message */
-    props.modalAction();
-    if (content !== 'modalContent') {
-      setInputIsValid('');
-      clearInputs();
+    const token = new URLSearchParams(window.location.search).get('token');
+    const { email } = jwtDecode(token);
+    const { password, confirmPassword } = this.state;
+    const validateInput = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{6,}$/;
+    if (!password.match(validateInput)) {
+      toast.error('Password must contain uppercase and lowercase letters, a number and not be less than 6 characters');
+    } else if (password === confirmPassword) {
+      // calling the action creator to update the store
+      const { resetPassword, history } = this.props;
+      await resetPassword(token, password, confirmPassword, email);
+      const { message } = this.props;
+      message === 'Password reset successful'
+        ? toast.success('Password updated successfully !')
+        : toast.error('Your request to reset password has expired or check your internet connection.');
+      history.push('/login');
+    } else {
+      toast.error('Password do not match');
     }
   };
 
-  return (
-    <Modal clearMessage={clearModalState} title="Recover Password!">
-      <p className="reset_password_paragraph">Enter the email address you used when you created your account. You will receive instructions to reset your password in your mail box.</p>
-      <form className={`reset-password-group ${inputIsValid}`} onSubmit={(event => handleSubmit(event))}>
-        <div className="form-group">
-          <input ref={input} type="email" autoComplete="true" className="reset_form_control btn-block" placeholder="Enter email" name="email" onChange={event => updateInput(event)} required />
-          <span className="input-danger mt-2">A valid email is required</span>
-          <button type="submit" className="reset_btn btn-block mt-3">Request reset password link</button>
-        </div>
-
-      </form>
-      <div>
-        {
-          isLoading
-          && (
-            <div className="reset_spinner_box">
-              <ClipLoader
-                size={40}
-                color="purple"
-              />
-            </div>
-          )
-        }
-      </div>
-      {
-        message && <div className="reset_alert_danger">{message}</div>
-      }
-    </Modal>
-  );
+  /**
+   * @memberof ResetPassword
+   * @returns {JSX} The JSX representation of the Reset Password component
+  */
+  render() {
+    return (
+      <ResetPasswordForm
+        handleSubmit={this.handleSubmit}
+        updateInput={this.updateInput}
+      />
+    );
+  }
 }
 
 ResetPassword.propTypes = {
+  resetPassword: PropTypes.func,
+  history: PropTypes.object,
   message: PropTypes.string,
-  isLoading: PropTypes.bool,
-  modalAction: PropTypes.func,
-  passwordResetRequest: PropTypes.func,
 };
 
 ResetPassword.defaultProps = {
-  message: '',
-  isLoading: false,
-  modalAction: f => f,
-  passwordResetRequest: f => f,
+  resetPassword: () => {},
 };
 
-const mapStateToProps = createStructuredSelector(
-  {
-    isLoading: selectors.getResetPassswordLoading,
-    message: selectors.getResetPasswordMessage,
-  }
-);
+const mapStateToProps = createStructuredSelector({
+  message: selectors.getResetPasswordMessage,
+});
 
-export default connect(mapStateToProps, { modalAction, passwordResetRequest })(ResetPassword);
+
+export default connect(mapStateToProps, { resetPassword })(ResetPassword);
