+++ src/reducers/resetPasswordReducer.js
@@ -1,4 +1,4 @@
-import { resetPasswordType } from '../actions/resetPassword';
+import { resetPasswordType } from '../actions/resetPasswordAction';
 import { resetModal } from '../actions/modal';
 
 const initialState = {
@@ -14,17 +14,44 @@ const initialState = {
  * @param {object} action - to update state
  * @returns {object}
  */
-
-export default function resetPasswordReducer(state = initialState, action = {}) {
+export default function resetPasswordReducer(
+  state = initialState,
+  action = {}
+) {
   const { type, message } = action;
   switch (type) {
     case resetPasswordType.loading:
-      return { ...state, resetPassword: { loading: true, message: '' } };
+      return {
+        ...state,
+        resetPassword: {
+          loading: true,
+          message: '',
+        }
+      };
     case resetPasswordType.success:
+      return {
+        ...state,
+        resetPassword: {
+          loading: false,
+          message,
+        }
+      };
     case resetPasswordType.failure:
-      return { ...state, resetPassword: { loading: false, message } };
+      return {
+        ...state,
+        resetPassword: {
+          loading: false,
+          message
+        }
+      };
     case resetModal.success:
-      return { ...state, resetPassword: { loading: false, message: false } };
+      return {
+        ...state,
+        resetPassword: {
+          loading: false,
+          message,
+        }
+      };
     default:
       return state;
   }
