+++ src/routes/index.jsx
@@ -10,7 +10,7 @@ import Navbar from '../components/NavbarContainer';
 import Article from '../components/Article';
 import NotFound from '../components/NotFound';
 import SearchContainer from '../components/SearchContainer';
-import UpdateUserPassword from '../components/UpdateUserPassword';
+import ResetPassword from '../components/ResetPassword';
 
 /**
  * @function Routes - A JSX wrapper for all the app's routes
@@ -24,7 +24,7 @@ function Routes() {
         <Navbar />
         <Switch>
           <Route path="/" component={ConnectedHomePage} exact />
-          <Route path="/reset-password" component={UpdateUserPassword} exact />
+          <Route path="/reset-password" component={ResetPassword} exact />
           <Route path="/login" component={ConnectedLogin} exact />
           <Route path="/signup" component={SignupPage} exact />
           <Route path="/search" component={SearchContainer} exact />
