+++ src/test/actions/resetPassAction.test.js
@@ -1,6 +1,14 @@
-import expect from 'expect';
-import * as actions from '../../actions/resetPassword';
+import configureMockStore from 'redux-mock-store';
+import thunk from 'redux-thunk';
+import * as actions from '../../actions/resetPasswordAction';
+// import { resetPassword } from '../../actions/resetPasswordAction';
+import mockAxios from '../../__mocks__/axios';
+import generateToken from '../utils/generateToken';
 
+const middlewares = [thunk];
+const mockStore = configureMockStore(middlewares);
+
+const token = generateToken({ email: 'alexandra.collins+2@andela.com', id: 1 });
 
 describe('Action creators', () => {
   it('should create an action to setLoading state', () => {
@@ -24,5 +32,41 @@ describe('Action creators', () => {
       message: 'failed',
     };
     expect(actions.resetPasswordFailure('failed')).toEqual(expected);
+  });
+});
+
+describe('resetPasswordAction test suite', () => {
+  const data = {
+    token,
+    password: 'password1',
+    confirmPassword: 'password1',
+    email: 'alexandra.collins+2@andela.com'
+  };
+  it('creates RESET_PASSWORD_SUCCESS after successfuly updating a user password', () => {
+    mockAxios.patch.mockImplementationOnce(() => Promise.resolve({
+      data: {
+        token,
+        password: 'password1',
+        confirmPassword: 'password1',
+        email: 'alexandra.collins+2@andela.com'
+      }
+    }));
+
+    const expectedActions = [
+      { type: actions.resetPasswordType.loading },
+      { message: undefined, type: actions.resetPasswordType.success },
+    ];
+
+    const store = mockStore();
+
+    return store.dispatch(actions
+      .resetPassword(
+        data.token,
+        data.password,
+        data.confirmPassword,
+        data.email
+      )).then(() => {
+      expect(store.getActions()).toEqual(expectedActions);
+    });
   });
 });
