+++ src/test/reducers/resetPassReducer.test.js
@@ -1,5 +1,6 @@
 import resetPasswordReducer from '../../reducers/resetPasswordReducer';
-import { resetPasswordType } from '../../actions/resetPassword';
+import { resetPasswordType } from '../../actions/resetPasswordAction';
+import { resetModal } from '../../actions/modal';
 import { mockState } from '../setup/index';
 
 const { user } = mockState;
@@ -21,6 +22,28 @@ describe('Reset password reducer: ', () => {
     };
     const result = resetPasswordReducer(user, action);
     const expected = { ...user, resetPassword: { loading: false, message: 'successful' } };
+    expect(result).toEqual(expected);
+  });
+
+  it('should update the reducer state when loading', () => {
+    const action = {
+      type: resetPasswordType.loading,
+      loading: true,
+    };
+    const result = resetPasswordReducer(user, action);
+    const expected = { ...user, resetPassword: { loading: true, message: '' } };
+    expect(result).toEqual(expected);
+  });
+
+
+  it('should update the reducer state when resetting modal', () => {
+    const action = {
+      type: resetModal.success,
+      loading: false,
+      message: 'reset modal',
+    };
+    const result = resetPasswordReducer(user, action);
+    const expected = { ...user, resetPassword: { loading: false, message: 'reset modal' } };
     expect(result).toEqual(expected);
   });
